+++ cmd/cipdip/main.go
@@ -31,9 +31,10 @@ CIP/EtherNet-IP traffic for firewall DPI research testing.`,
 	rootCmd.AddCommand(newInstallCmd())
 	rootCmd.AddCommand(newPcapCmd())
 	rootCmd.AddCommand(newPcapSummaryCmd())
+	rootCmd.AddCommand(newPcapReportCmd())
+	rootCmd.AddCommand(newPcapClassifyCmd())
 	rootCmd.AddCommand(newTestCmd())
 	rootCmd.AddCommand(newBaselineCmd())
-	rootCmd.AddCommand(newExtractReferenceCmd())
 	rootCmd.AddCommand(newExtractReferenceCmd())
 
 	// Let Cobra handle help automatically - it will show flags and detailed descriptions